refactor(1413): use global UpdateProfileForm
