--- src/features/student/user/components/overlays/UpdateProfileDrawer/UpdateProfileDrawer.stub.ts
@@ -1,5 +1,0 @@
-export const UpdateProfileDrawerStub = defineComponent({
-  name: 'UpdateProfileDrawer',
-  props: ['show'],
-  template: '<div class="update-profile-drawer" />'
-})
