--- src/features/student/user/components/overlays/UpdateProfileDrawer/config.ts
@@ -1,1 +1,0 @@
-export const BIOGRAPHY_MAX_LENGTH = 400
