--- src/features/student/user/components/overlays/UpdateProfileDrawer/use-update-profile-form.ts
@@ -1,113 +1,0 @@
-import type { ProfileOverviewDTO } from '@/api/avenir-esr'
-import type { BaseApiException } from '@/common/exceptions'
-import { BIOGRAPHY_MAX_LENGTH } from '@/features/student/user/components/overlays/UpdateProfileDrawer/config'
-import { useUpdateProfile, useUpdateProfileCover, useUpdateProfilePhoto } from '@/features/student/user/components/overlays/UpdateProfileDrawer/use-update-profile'
-import { useToasterStore } from '@/store'
-import { isValidEmail } from '@avenirs-esr/avenirs-dsav'
-import { useForm, useStore } from '@tanstack/vue-form'
-import { useI18n } from 'vue-i18n'
-
-export function useUpdateProfileForm (initialData: ProfileOverviewDTO, onSuccess: () => void) {
-  const { t } = useI18n()
-  const { addErrorMessage } = useToasterStore()
-  const { onUpdateProfile, isUpdateProfilePending } = useUpdateProfile(onSuccess)
-  const { onUpdateProfileCoverAsync, isUpdateProfileCoverPending } = useUpdateProfileCover(onUpdateProfileCoverSuccess)
-  const { onUpdateProfilePhotoAsync, isUpdateProfilePhotoPending } = useUpdateProfilePhoto(onUpdateProfilePhotoSuccess)
-
-  const coverPictureFile = ref<File | null>(null)
-  const profilePictureFile = ref<File | null>(null)
-
-  const validateBiography = (value: string) => {
-    if (!value) {
-      return
-    }
-    return value.trim().length > BIOGRAPHY_MAX_LENGTH ? t('student.user.overlays.UpdateProfileDrawer.identity.biography.errors.tooLong', { maxLength: BIOGRAPHY_MAX_LENGTH }) : undefined
-  }
-  const form = useForm({
-    defaultValues: { ...initialData },
-    validators: {
-      onSubmit ({ value }) {
-        return {
-          fields: {
-            email: value.email && !isValidEmail(value.email) ? t('global.error.form.invalidEmail') : undefined,
-            bio: validateBiography(value.bio)
-          }
-        }
-      }
-    },
-    onSubmit: async ({ value }) => {
-      try {
-        const { email, bio } = value
-
-        const [coverResult, photoResult] = await Promise.allSettled([
-          coverPictureFile.value ? onUpdateProfileCoverAsync({ file: coverPictureFile.value }) : Promise.resolve(),
-          profilePictureFile.value ? onUpdateProfilePhotoAsync({ file: profilePictureFile.value }) : Promise.resolve()
-        ])
-
-        if (coverResult.status === 'rejected') {
-          throw coverResult.reason
-        }
-
-        if (photoResult.status === 'rejected') {
-          throw photoResult.reason
-        }
-
-        onUpdateProfile({ email, bio })
-      }
-      catch (error) {
-        addErrorMessage({
-          title: t('student.user.overlays.UpdateProfileDrawer.onUpdate.error'),
-          description: (error as BaseApiException)?.message ?? t('global.error.generic')
-        })
-      }
-    },
-  })
-
-  function onUpdateProfileCoverSuccess (url: string) {
-    form.setFieldValue('coverPicture.url', url)
-  }
-
-  function onUpdateProfilePhotoSuccess (url: string) {
-    form.setFieldValue('profilePicture.url', url)
-  }
-
-  function onCoverPictureUpdate (file: File | null) {
-    coverPictureFile.value = file
-  }
-
-  function onProfilePictureUpdate (file: File | null) {
-    profilePictureFile.value = file
-  }
-
-  const isPending = computed(() =>
-    isUpdateProfilePending.value
-    || isUpdateProfileCoverPending.value
-    || isUpdateProfilePhotoPending.value
-    || form.state.isSubmitting
-  )
-
-  const isDefaultValue = useStore(form.store, state => state.isDefaultValue)
-
-  const isModified = computed(() =>
-    !isDefaultValue.value || !!coverPictureFile.value || !!profilePictureFile.value
-  )
-
-  function resetForm () {
-    form.reset(initialData)
-    coverPictureFile.value = null
-    profilePictureFile.value = null
-  }
-
-  return {
-    form,
-    isPending,
-    isModified,
-    resetForm,
-    coverPictureFile,
-    onCoverPictureUpdate,
-    profilePictureFile,
-    onProfilePictureUpdate,
-    onUpdateProfileCoverSuccess,
-    onUpdateProfilePhotoSuccess
-  }
-}
