+++ src/features/student/user/index.ts
@@ -2,6 +2,5 @@ export { default as StudentOverviewWidget } from '@/features/student/user/compon
 export { default as StudentMailboxPopover } from '@/features/student/user/components/overlays/StudentMailboxPopover/StudentMailboxPopover.vue'
 export { default as StudentNotificationsPopover } from '@/features/student/user/components/overlays/StudentNotificationsPopover/StudentNotificationsPopover.vue'
 export { default as StudentProfilePopover } from '@/features/student/user/components/overlays/StudentProfilePopover/StudentProfilePopover.vue'
-export { default as UpdateProfileDrawer } from '@/features/student/user/components/overlays/UpdateProfileDrawer/UpdateProfileDrawer.vue'
 export { useStudentNavigationAccessControlQuery } from '@/features/student/user/queries/use-student-access-control.query/use-student-access-control.query'
 export { useStudentUserStore } from '@/features/student/user/stores/user.store'
